Beacon distance determined to 1m, fixed lifecycle bug, added progress bar -> set to 10seconds, added wallet dialog, added location markers around the university campus

+++ app/src/main/java/com/project/level4/adgo/utils/Wallet.java
@@ -16,7 +16,7 @@ public class Wallet {
         return balance;
     }
 
-    public void addToWallet(long value){
+    public void addToWallet(double value){
         balance += value;
     }
